+++ src/ipl3checksum/__init__.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 
-# SPDX-FileCopyrightText: © 2023 Decompollaborate
+# SPDX-FileCopyrightText: © 2023-2024 Decompollaborate
 # SPDX-License-Identifier: MIT
 
 from __future__ import annotations
 
 # Version should be synced with pyproject.toml, Cargo.toml and src/rs/version.rs
-__version_info__: tuple[int, int, int] = (1, 1, 1)
+__version_info__: tuple[int, int, int] = (1, 2, 0)
 __version__ = ".".join(map(str, __version_info__))
 __author__ = "Decompollaborate"